change api route error handling

+++ app/api/user/delete/route.ts
@@ -6,6 +6,12 @@ export async function DELETE(req: NextRequest, res: NextResponse) {
   try {
     const body = await req.json();
 
+    if (!body) {
+      return NextResponse.json("Bad Request - Missing required parameters.", {
+        status: 400,
+      });
+    }
+
     const user = await prisma.user.delete({
       where: {
         email: body,
@@ -14,7 +20,7 @@ export async function DELETE(req: NextRequest, res: NextResponse) {
 
     return NextResponse.json(user);
   } catch (error) {
-    console.error("[REGISTER_POST]", error);
+    console.error("REGISTER_POST", error);
     return new NextResponse("Internal Error", { status: 500 });
   }
 }
